types file for structs and enums
